Introducing: a simple compiler
Previously, words were executed (or skipped, e.g. one branch of an
if/else) as soon as they were encountered in the source code. This doesn't
allow us to implement user-defined words because there's no way to delay
the execution of a word.

In order to enable user-defined words I implemented a compiler. For
simplicity's sake every word is now compiled. Compiling a word returns an
execution token (xt) which is executed at the right time.

User-defined word implementation coming next.